Split out internal types.
