+++ packages/gasket-plugin-happyfeet/lib/index.d.ts
@@ -2,11 +2,10 @@ import type { HappyFeet, HappyFeetOptions } from 'happy-feet';
 import type { MaybeAsync } from '@gasket/engine';
 
 declare module '@gasket/engine' {
-    export interface GasketConfig {
-        happyFeet: HappyFeetOptions
-    }
-    export interface Gasket {
-        happyFeet: HappyFeet;
-    }
+  export interface GasketConfig {
+    happyFeet: HappyFeetOptions;
+  }
+  export interface Gasket {
+    happyFeet: HappyFeet;
+  }
 }
-
