+++ packages/gasket-plugin-intl/lib/index.d.ts
@@ -67,18 +67,3 @@ declare module '@gasket/cli' {
     hasGasketIntl?: boolean;
   }
 }
-
-/** Tuple of package name and path */
-type SrcPkgDir = [string, string];
-
-async function getPackageDirs(
-  /** Path to parent directory */
-  parentDir: string,
-  /** List of full paths */
-  dirList?: SrcPkgDir[]
-): AsyncGenerator<SrcPkgDir>;
-
-function withLocaleRequired(
-  /** Path(s) containing locale files */
-  localePathPart: LocalePathPart | LocalePathPart[]
-): LocalesProps;
