+++ packages/gasket-plugin-intl/lib/internal.d.ts
@@ -12,3 +12,18 @@ function formatLocale(
   /** Selected accept-language */
   language: string
 ): string;
+
+/** Tuple of package name and path */
+type SrcPkgDir = [string, string];
+
+async function getPackageDirs(
+  /** Path to parent directory */
+  parentDir: string,
+  /** List of full paths */
+  dirList?: SrcPkgDir[]
+): AsyncGenerator<SrcPkgDir>;
+
+function withLocaleRequired(
+  /** Path(s) containing locale files */
+  localePathPart: LocalePathPart | LocalePathPart[]
+): LocalesProps;
