+++ packages/gasket-plugin-lint/lib/index.d.ts
@@ -33,18 +33,3 @@ declare module '@gasket/cli' {
     standard?: StandardConfig;
   }
 }
-
-export interface CodeStyle {
-  name?: string;
-  allowStylelint?: boolean;
-  create?: (context: CreateContext, utils: Utils) => Promise<void>;
-}
-
-export interface Utils {
-  gatherDevDeps(pkgName: string): Promise<Record<string, string>>;
-  runScriptStr(script: string): string;
-}
-
-export interface LintUtils {
-  makeGatherDevDeps(context: CreateContext): Promise<string[]>;
-}
