+++ packages/gasket-plugin-lint/lib/internal.d.ts
@@ -1,0 +1,16 @@
+import type { CreateContext } from '@gasket/cli';
+
+export interface Utils {
+  gatherDevDeps(pkgName: string): Promise<Record<string, string>>;
+  runScriptStr(script: string): string;
+}
+
+export interface CodeStyle {
+  name?: string;
+  allowStylelint?: boolean;
+  create?: (context: CreateContext, utils: Utils) => Promise<void>;
+}
+
+export interface LintUtils {
+  makeGatherDevDeps(context: CreateContext): Promise<string[]>;
+}
