+++ packages/gasket-plugin-manifest/lib/internal.d.ts
@@ -1,0 +1,11 @@
+import { Gasket } from '@gasket/engine';
+
+export function gatherManifestData(
+  gasket: Gasket,
+  context: {
+    req?: IncomingMessage & {
+      originalUrl?: string;
+    };
+    res?: OutgoingMessage;
+  }
+): MaybeAsync<Manifest>;
