+++ packages/gasket-plugin-morgan/lib/index.d.ts
@@ -14,7 +14,13 @@ declare module '@gasket/engine' {
 export declare const dependencies: string[];
 
 export declare namespace hooks {
-    namespace middleware {
-        function handler(gasket: GasketConfig): ((req: IncomingMessage, res: ServerResponse, next: (err?: Error) => void) => void)[];
-    }
+  namespace middleware {
+    function handler(
+      gasket: GasketConfig
+    ): ((
+      req: IncomingMessage,
+      res: ServerResponse,
+      next: (err?: Error) => void
+    ) => void)[];
+  }
 }
