+++ packages/gasket-plugin-morgan/lib/index.js
@@ -1,6 +1,7 @@
-const morgan = require('morgan');
-const split = require('split');
+/// <reference types="@gasket/plugin-metadata" />
+
 const { name } = require('../package.json');
+const middleware = require('./middleware');
 
 // set log configuration in gasket.config.js, under `morgan` key
 // configuration options: http://expressjs.com/en/resources/middleware/morgan.html
@@ -9,38 +10,31 @@ const plugin = {
   name,
   dependencies: ['@gasket/plugin-log'],
   hooks: {
-    middleware: {
-      handler: (gasket) => {
-        const { logger, config } = gasket;
-        const { morgan: { format = 'tiny', options = {} } = {} } = config;
-
-        const stream = split().on('data', (line) => logger.info(line));
-
-        const morganMiddleware = morgan(format, { ...options, stream });
-
-        return [morganMiddleware];
-      }
-    },
+    middleware,
     metadata(gasket, meta) {
       return {
         ...meta,
-        configurations: [{
-          name: 'morgan',
-          link: 'README.md#configuration',
-          description: 'Morgan plugin configuration',
-          type: 'object'
-        }, {
-          name: 'morgan.format',
-          link: 'README.md#configuration',
-          description: 'Log format to print',
-          type: 'string',
-          default: 'tiny'
-        }, {
-          name: 'morgan.options',
-          link: 'README.md#configuration',
-          description: 'Morgan options',
-          type: 'object'
-        }]
+        configurations: [
+          {
+            name: 'morgan',
+            link: 'README.md#configuration',
+            description: 'Morgan plugin configuration',
+            type: 'object'
+          },
+          {
+            name: 'morgan.format',
+            link: 'README.md#configuration',
+            description: 'Log format to print',
+            type: 'string',
+            default: 'tiny'
+          },
+          {
+            name: 'morgan.options',
+            link: 'README.md#configuration',
+            description: 'Morgan options',
+            type: 'object'
+          }
+        ]
       };
     }
   }
