+++ packages/gasket-plugin-nextjs/lib/index.d.ts
@@ -81,17 +81,3 @@ declare module '@gasket/plugin-nextjs' {
     return Promise.resolve(null);
   }
 }
-
-/**
- * Small helper function that creates nextjs app from the gasket configuration.
- */
-export async function setupNextApp(gasket: Gasket): NextServer;
-
-/**
- * Sets up the next.js request handler to be called after all other middleware
- */
-export function setupNextServer(
-  nextServer: NextServer,
-  serverApp: Application | Fastify,
-  gasket: Gasket
-): void;
